Create README.md
This is going to be a large scale project concerned with quantifying astrocyte networks and individual cellular morphology.
As of now there is a series of scripts that utilizes the breadth first search algorithm to traverse a binary image between two nodes to see if they are directly connected. 
Next step is to get 100x tile confocal stack of a tdTomato labeled astrocyte mouse retina and determine how many other astrocytes each cell contacts. This will be plotted as a heat map to determine if there are regional differences with distance from the optic nerve head and also between major blood vessel rich/poor regions.